+++ tests/cases/unplot/plan.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+from stepup.core.api import static
+from stepup.reprep.api import unplot
+
+static("plot.svg")
+unplot("plot.svg", "current_plot.json")